Automatically delete PVCs on pipeline completion
By adding tekton.dev/pvc-protection-finalizer: remove to the
metadata.labels of the volumeClaimTemplate, we instruct the Tekton
controller to delete the PVC once the PipelineRun completes its
execution. It has been noticed that with the later label, if the PVC is
not deleted on PipelineRun completion, it is deleted when the
PipelineRun is deleted.